refactor: [Frontend] moved components of NewTestRunPanel into separate folder
