+++ frontend/src/components/testrun/newtestrun/NewTestRunStoarage.tsx
@@ -1,0 +1,18 @@
+import {convertNewTestRunFromDTO, convertNewTestRunIntoDTO, NewTestRun} from "../Types.tsx";
+
+const KEY_CURRENT_NEW_TEST_RUN: string = "NewTestRunPanel.CurrentNewTestRun";
+
+export function saveCurrentNewTestRun( scenarioId: string, newTestRun: NewTestRun ) {
+    localStorage.setItem(KEY_CURRENT_NEW_TEST_RUN+"."+scenarioId, JSON.stringify( convertNewTestRunIntoDTO(newTestRun) ));
+}
+export function loadCurrentNewTestRun(scenarioId: string): NewTestRun | undefined  {
+    const str = localStorage.getItem(KEY_CURRENT_NEW_TEST_RUN+"."+scenarioId);
+    if (str) return convertNewTestRunFromDTO( JSON.parse(str) );
+}
+export function isCurrentNewTestRunStored(scenarioId: string): boolean  {
+    const str = localStorage.getItem(KEY_CURRENT_NEW_TEST_RUN+"."+scenarioId);
+    return !(!str);
+}
+export function clearCurrentNewTestRun(scenarioId: string) {
+    localStorage.setItem(KEY_CURRENT_NEW_TEST_RUN+"."+scenarioId, "");
+}
