+++ frontend/src/components/testrun/newtestrun/TestCasesView.tsx
@@ -1,8 +1,8 @@
-import {Id, Label} from "../StandardStyledComponents.tsx";
-import {compareStringsIgnoringCase} from "../../Tools.tsx";
+import {Id, Label} from "../../StandardStyledComponents.tsx";
+import {compareStringsIgnoringCase} from "../../../Tools.tsx";
 import styled from "styled-components";
-import {TestCase} from "./Types.tsx";
-import {SHOW_RENDERING_HINTS} from "../../Types.tsx";
+import {TestCase} from "../Types.tsx";
+import {SHOW_RENDERING_HINTS} from "../../../Types.tsx";
 
 const SimpleCard = styled.div`
   border: 1px solid var(--border-color, #707070);
